+++ lib/src/validators/number_validator_error.dart
@@ -1,0 +1,6 @@
+class NumberValidatorError {
+  static const String nullValue = 'nullValue';
+  static const String invalidDecimals = 'invalidDecimals';
+  static const String unsignedNumber = 'unsignedNumber';
+  static const String invalidNumber = 'invalidNumber';
+}