+++ lib/src/widgets/reactive_slider.dart
@@ -31,9 +31,9 @@ class ReactiveSlider extends ReactiveFocusableFormField<num, double> {
   /// The [labelBuilder] is called each time the [FormControl] changes its value
   /// so you can supply a label to the Slider.
   ReactiveSlider(
-      {Key? key,
-      String? formControlName,
-      FormControl<num>? formControl,
+      {super.key,
+      super.formControlName,
+      super.formControl,
       double min = 0.0,
       double max = 1.0,
       int? divisions,
@@ -44,7 +44,7 @@ class ReactiveSlider extends ReactiveFocusableFormField<num, double> {
       SemanticFormatterCallback? semanticFormatterCallback,
       bool autofocus = false,
       MouseCursor? mouseCursor,
-      FocusNode? focusNode,
+      super.focusNode,
       ReactiveFormFieldCallback<num>? onChangeEnd,
       ReactiveFormFieldCallback<num>? onChangeStart,
       ReactiveFormFieldCallback<num>? onChanged,
@@ -52,10 +52,6 @@ class ReactiveSlider extends ReactiveFocusableFormField<num, double> {
       Color? secondaryActiveColor,
       MaterialStateProperty<Color?>? overlayColor})
       : super(
-          key: key,
-          formControl: formControl,
-          formControlName: formControlName,
-          focusNode: focusNode,
           builder: (field) {
             var value = field.value;
             if (value == null) {