+++ lib/src/widgets/reactive_text_field.dart
@@ -90,13 +90,13 @@ class ReactiveTextField<T> extends ReactiveFormField<T, String> {
   /// For documentation about the various parameters, see the [TextField] class
   /// and [TextField], the constructor.
   ReactiveTextField({
-    Key? key,
-    String? formControlName,
-    FormControl<T>? formControl,
-    Map<String, ValidationMessageFunction>? validationMessages,
-    ControlValueAccessor<T, String>? valueAccessor,
-    ShowErrorsFunction<T>? showErrors,
-    FocusNode? focusNode,
+    super.key,
+    super.formControlName,
+    super.formControl,
+    super.validationMessages,
+    super.valueAccessor,
+    super.showErrors,
+    super.focusNode,
     InputDecoration decoration = const InputDecoration(),
     TextInputType? keyboardType,
     TextCapitalization textCapitalization = TextCapitalization.none,
@@ -158,13 +158,6 @@ class ReactiveTextField<T> extends ReactiveFormField<T, String> {
     TextMagnifierConfiguration? magnifierConfiguration,
   })  : _textController = controller,
         super(
-          key: key,
-          formControl: formControl,
-          formControlName: formControlName,
-          valueAccessor: valueAccessor,
-          validationMessages: validationMessages,
-          showErrors: showErrors,
-          focusNode: focusNode,
           builder: (ReactiveFormFieldState<T, String> field) {
             final state = field as _ReactiveTextFieldState<T>;
             final effectiveDecoration = decoration
@@ -302,9 +295,10 @@ class _ReactiveTextFieldState<T>
 
   @override
   void dispose() {
-    if (widget.textController == null) {
+    final currentWidget = widget as ReactiveTextField<T>;
+    if (currentWidget._textController == null) {
       _textController.dispose();
-    }    
+    }
     super.dispose();
   }
 }